requirements.txt
fastapi
uvicorn
python-multipart
jinja2
sqlalchemy
python-decouple
supabase-py